Changing the dashboard for frontend and adding user creation in backend

+++ backend/routes/api.php
@@ -13,28 +13,46 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::middleware(['jwt.auth'])->get('/me','HomeController@me');
+Route::middleware(['jwt.auth'])->get('/me', 'HomeController@me');
 
-Route::middleware(['jwt.auth'])->prefix('application')->group(static function () {
-    Route::get('/', 'ApplicationController@getApplications');
-    Route::get('/{id}', 'ApplicationController@getApplication');
-    Route::post('/', 'ApplicationController@createApplication');
-    Route::post('/app-key', 'ApplicationController@createAppKey');
-    Route::delete('/{id}', 'ApplicationController@deleteApplication');
-    Route::delete('/app-key/{id}', 'ApplicationController@deleteAppKey');
-});
+Route::middleware(['jwt.auth'])
+    ->prefix('application')
+    ->group(
+        static function () {
+            Route::get('/', 'ApplicationController@getApplications');
+            Route::get('/{id}', 'ApplicationController@getApplication');
+            Route::post('/', 'ApplicationController@createApplication');
+            Route::post('/app-key', 'ApplicationController@createAppKey');
+            Route::delete('/{id}', 'ApplicationController@deleteApplication');
+            Route::delete('/app-key/{id}', 'ApplicationController@deleteAppKey');
+        }
+    );
 
-Route::middleware('jwt.auth')->prefix('users')->group(static function () {
-    Route::post('/', 'AuthApi\UserController@create');
-    Route::delete('/{id}', 'AuthApi\UserController@delete');
-    Route::post('/change-image', 'AuthApi\UserController@changeImage');
-});
+Route::middleware('jwt.auth')
+    ->prefix('users')
+    ->group(
+        static function () {
+            Route::middleware('permission:create-users')->post('/', 'AuthApi\UserController@create');
+            Route::middleware(['permission:delete-user'])->delete('/{id}', 'AuthApi\UserController@delete');
+            Route::post('/change-image', 'AuthApi\UserController@changeImage');
+        }
+    );
 
-Route::middleware(['jwt.auth'])->prefix('messages')->group(static function () {
-    Route::post('/{applicationId}', 'MessageController@createMessage');
-});
-Route::middleware(['app_key'])->prefix('subscribers')->group(static function () {
-    Route::post('/', 'SubscriptionController@subscribe');
-});
+Route::middleware(['jwt.auth'])
+    ->prefix('messages')
+    ->group(
+        static function () {
+            Route::post('/{applicationId}', 'MessageController@createMessage');
+        }
+    );
+
+
+Route::middleware(['app_key'])
+    ->prefix('subscribers')
+    ->group(
+        static function () {
+            Route::post('/', 'SubscriptionController@subscribe');
+        }
+    );
